Docker, Dockerhub, docker-compose and script.sh READ DESCRIPTION
FIRST, DONT TOUCH ANYTHING.
THEN, you need to stop MySQL server of your computer, because he is using port 3306, and we need that port on docker-compose to execute the database, so please before all, stop the MySQL process.
-You should know that you need to get docker desktop for your computer.
-Then, you need to "docker pull lordkener/instarecipes", to get the repository of the project.
-Then, for run, there are 2 possibilities:
If youre working on windows, you maybe should clone the script.sh and put it like "script.ps1", where .ps1 extension is the PowerShell of windows. Maybe you dont need to do this and simply just execute the .sh, just try it. But DONT CHANGE ANYTHING.
-Maybe you just need to run .sh FROM THE CMD, or Shell in linux or mac.